Initial Commit
I only worked on this for about 30 minutes, but I got the basic systems in place. I can go up to three fruits.

I am using the fruit name to determine if they are the same and an age system to determine which would should process the merge. Right now the new fruit spawns in at the midpoint between the two objects. However it might be smarter to do it at the Collision point.

I'd also like to actually add sprites. I'm not going to sell this game, however I would like it to look okay.