Adding the death plane.
I need to think about what is going to properly handle the end game. Right now I've got it that the MouseTracker has a boolean that controls whether or not a piece can be placed. If that boolean is false, the game is over, you've died. But it's not really a Game Manager. I might want to upgrade the ScoreContainer to become a game manager since score would be an aspect of the game that would need to be managed.